refactor: consistent module naming
